+++ src/gevent/tests/monkey_package/__init__.py
@@ -2,6 +2,10 @@
 """
 Make a package.
 
+This file has no other functionality. Individual modules in this package
+are used for testing, often being run with 'python -m ...' in individual
+test cases (functions).
+
 """
 from __future__ import absolute_import
 from __future__ import division